app/requirements.txt
pandas
psycopg2-binary
numpy
pycpfcnpj